+++ docs/src/news.md
@@ -38,6 +38,7 @@ multithreading, resource handling and a streamlined documentation.
 - actors can register their message channels to the `clock.channels` vector and the clock will not proceed before they are empty,
 - processes and actors (asynchronous tasks) can transfer IO-operations to the clock with [`now!`](@ref) or print directly via the clock,
 - `event!` and `delay!` now also accept stochastic time variables (`Distribution`).
+- there is a `n` keyword parameter for repeat `event!`s.
 
 ### Multithreading (experimental)